Slightly optimize memory usage
Save some memory by using smaller arrays for the even and odd terms.
Adjust the loop lengths, to accomodate for shorter arrays
not overwriting "outdated" data.

E.g.: If a and b are aliased, then the first poly_mul does not overwrite
the leading coefficient, which leads to incorrect results.