Entire (generalised) Bernoulli function
In Peter Luschny's "An introduction to the Bernoulli function"
(https://arxiv.org/abs/2009.06743) a generalised Bernoulli function is
defined:

          ⎧      1         for s = 0
B(s, a) = ⎨
          ⎩-s⋅ζ(1 - s, a)  otherwise

When s is a nonnegative integer this function reproduces the Bernoulli
polynomials; when in addition a = 1 (the default choice for the Hurwitz
zeta function's second argument, yielding Riemann's zeta) the Bernoulli
numbers are obtained. Hence B(1) = +1/2, which is how Jakob Bernoulli
originally defined his numbers in Ars Conjectandi (1713); this choice
gives the simpler form/wider range of validity than B(1) = -1/2 for very
many equations like Faulhaber's formula and the Euler–Maclaurin
summation formula.